pyproject.toml
[project]
name = "agenticflow"
version = "0.1.0"
description = "A production-grade event-driven multi-agent system framework"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
authors = [{ name = "Milad Olad" }]
keywords = ["agents", "multi-agent", "event-driven", "llm", "ai", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langgraph>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
websocket = ["websockets>=12.0"]
api = ["fastapi>=0.115.0", "uvicorn>=0.32.0"]
all = ["websockets>=12.0", "fastapi>=0.115.0", "uvicorn>=0.32.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[project.scripts]
agenticflow = "agenticflow.cli:main"

[project.urls]
Homepage = "https://github.com/milad-o/agenticflow"
Repository = "https://github.com/milad-o/agenticflow"
Issues = "https://github.com/milad-o/agenticflow/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agenticflow"]

[tool.hatch.build.targets.sdist]
include = ["src/agenticflow", "README.md", "LICENSE"]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["agenticflow"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
